refactor(ingest): split the raster swap block out of the replace task
fix(#1290 review): `tasks_raster_replace.py` was at 986 lines and round 4's
fixes cross the 1000-line ratchet threshold, so the split lands FIRST and on
its own — the fixes go on top of a tree that is already the right shape, which
is the clean history round 3 could not sequence.

A pure extraction. Every function moved verbatim; no rewrites are mixed in.

The seam is the one the task is already built around. `tasks_raster_replace`
keeps the pipeline — claim, validate, convert, verify, and the terminal cleanup
that decides what the uploaded bytes were worth. `tasks_raster_swap` takes what
happens once those bytes have become a COG worth publishing: the field swap,
the asset-row upsert, the superseded-object reap and the post-commit
follow-ups. One module produces an artifact, the other makes the catalog point
at it. 986 -> 743, and the new module is 274.

One test moves with it: `invalidate_catalog_cache` is called from
`_run_post_swap_followups`, so the post-commit-failure test's monkeypatch
target moved to the new module. Found by grepping the module-qualified form
across tests before running anything, which is the sweep the previous split
skipped and paid for.
